+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/ChannelUtils.kt
@@ -3,33 +3,15 @@ package io.getstream.chat.android.compose.ui.util
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import com.getstream.sdk.chat.viewmodel.messages.getCreatedAtOrThrow
-import io.getstream.chat.android.client.extensions.getUsersExcludingCurrent
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import java.util.Date
-
-/**
- * Generates the display name for a channel based on its data.
- *
- * @return The display name of the channel.
- */
-@Composable
-@ReadOnlyComposable
-public fun Channel.getDisplayName(): String {
-    return name.takeIf { it.isNotEmpty() }
-        ?: getUsersExcludingCurrent()
-            .joinToString { it.name }
-            .takeIf { it.isNotEmpty() }
-        ?: stringResource(id = R.string.stream_compose_untitled_channel)
-}
 
 /**
  * Returns channel's last regular or system message if exists.